Created a new class to manage SpiSim batch computation
