+++ pyaedt/circuit.py
@@ -8,16 +8,13 @@ import math
 import os
 import re
 import shutil
-import subprocess  # nosec B404
 
 from pyaedt import Hfss3dLayout
-from pyaedt import settings
 from pyaedt.application.AnalysisNexxim import FieldAnalysisCircuit
 from pyaedt.generic import ibis_reader
 from pyaedt.generic.DataHandlers import from_rkm_to_aedt
 from pyaedt.generic.filesystem import search_files
 from pyaedt.generic.general_methods import generate_unique_name
-from pyaedt.generic.general_methods import is_linux
 from pyaedt.generic.general_methods import open_file
 from pyaedt.generic.general_methods import pyaedt_function_handler
 from pyaedt.modules.Boundary import CurrentSinSource
@@ -1687,266 +1684,3 @@ class Circuit(FieldAnalysisCircuit, object):
         hfss_3d_layout_model = self.modeler.schematic.add_subcircuit_3dlayout(hfss.design_name)
         hfss.close_project(save_project=False)
         return hfss_3d_layout_model
-
-    @pyaedt_function_handler()
-    def compute_erl_from_touchstone(
-        self,
-        touchstone_file,
-        config_file=None,
-        port_order=None,
-        specify_through_ports=None,
-        bandwidth=None,
-        tdr_duration=None,
-        z_terminations=None,
-        transition_time=None,
-        fixture_delay=None,
-        input_amplitude=None,
-        ber=None,
-        pdf_bin_size=None,
-        signal_loss_factor=None,
-        permitted_reflection=None,
-        reflections_lenght=None,
-        modulation_type=None,
-    ):
-        """Compute effective return loss (erl) using Ansys SPISIM from a S-parameter file.
-
-        Parameters
-        ----------
-        touchstone_file : str
-            Path to the touchstone file.
-        config_file : str, optional
-            Configuration file to use as a reference. The default is ``None``, in
-            which case this parameter is ignored.
-        port_order : str, optional
-            Whether to use "``EvenOdd``" or "``Incremental``" numbering for ``s4p`` files.
-            Ignored if the ports are greater than 4.
-        specify_through_ports : list, optional
-            Input and Output ports on which compute the erl. Those are ordered like ``[inp, inneg, outp, outneg]``.
-            Ignored if number of ports are 4.
-        bandwidth : float, str, optional
-            Application bandwidth: inverse of one UI (unit interval). Can be a float or str with unit ("m", "g").
-            Default is ``30e9``.
-        tdr_duration : float, optional
-            TDR duration (in second): How long the TDR tailed data should be applied. Default is ``5``.
-        z_terminations : float, optional
-            Z-Terminations: termination (Z11 and Z22) when TDR is calculated. Default is ``50``.
-        transition_time : float, str, optional
-            Transition time: how fast (slew rate) input pulse transit from 0 to Vcc volt. Default is "``10p``".
-        fixture_delay : float, optional
-            Fixture delay: delay when input starts transition from 0 to Vcc. Default is ``500e-12``.
-        input_amplitude : float, optional
-            Input amplitude: Vcc volt of step input. Default is ``1.0``.
-        ber : float, optional
-            Specified BER: At what threshold ERL is calculated. Default is ``1e-4``.
-        pdf_bin_size : float, optional
-            PDF bin size: how to quantize the superimposed value. Default is ``1e-5``.
-        signal_loss_factor : float, optional
-            Signal loss factor (Beta). See SPISIM Help for info. Default is ``1.7e9``.
-        permitted_reflection : float, optional
-            Permitted reflection (Rho). See SPISIM Help for info. Default is ``0.18``.
-        reflections_lenght : float, optional
-            Length of the reflections: how many UI will be used to calculate ERL. Default is ``1000``.
-        modulation_type : str, optional
-           Modulations type: signal modulation type "``NRZ``" or "``PAM4``". Default is "``NRZ``".
-
-        Returns
-        -------
-        bool or dict
-            A dictionary with the result from the spisimExe command, ``False`` when failed.
-        """
-        exec_name = "SPISimJNI_LX64.exe" if is_linux else "SPISimJNI_WIN64.exe"
-        spisimExe = os.path.join(self.desktop_install_dir, "spisim", "SPISim", "modules", "ext", exec_name)
-        cfg_dict = {
-            "INPARRY": "",
-            "MIXMODE": "",
-            "THRUS4P": "",
-            "BANDWID": 30e9,
-            "TDR_DUR": 5,
-            "BINSIZE": 1e-5,
-            "REFIMPD": 50,
-            "SPECBER": 1e-4,
-            "MODTYPE": "NRZ",
-            "FIXDELY": 500e-12,
-            "INPVOLT": 1.0,
-            "TRSTIME": "10p",
-            "SIGBETA": 1.7e9,
-            "REFLRHO": 0.18,
-            "NCYCLES": 1000,
-        }
-        if config_file:
-            with open(config_file, "r") as fp:
-                lines = fp.readlines()
-                for line in lines:
-                    if not line.startswith("#") and "=" in line:
-                        split_line = [i.strip() for i in line.split("=")]
-                        cfg_dict[split_line[0]] = split_line[1]
-        cfg_dict["INPARRY"] = touchstone_file
-        cfg_dict["MIXMODE"] = "" if "MIXMODE" not in cfg_dict else cfg_dict["MIXMODE"]
-        if port_order is not None and touchstone_file.lower().endswith(".s4p"):
-            cfg_dict["MIXMODE"] = port_order
-        elif not touchstone_file.lower().endswith(".s4p"):
-            cfg_dict["MIXMODE"] = ""
-        cfg_dict["THRUS4P"] = "" if "THRUS4P" not in cfg_dict else cfg_dict["THRUS4P"]
-
-        if specify_through_ports:
-            if isinstance(specify_through_ports[0], int):
-                thrus4p = ",".join([str(i) for i in specify_through_ports])
-            else:
-                try:
-                    ports = list(self.excitations.keys())
-                    thrus4p = ",".join([str(ports.index(i)) for i in specify_through_ports])
-                except IndexError:
-                    self.logger.error("Port not found.")
-                    return False
-            cfg_dict["THRUS4P"] = thrus4p
-
-        cfg_dict["BANDWID"] = bandwidth if bandwidth is not None else cfg_dict["BANDWID"]
-        cfg_dict["TDR_DUR"] = tdr_duration if tdr_duration is not None else cfg_dict["TDR_DUR"]
-        cfg_dict["BINSIZE"] = pdf_bin_size if pdf_bin_size is not None else cfg_dict["BINSIZE"]
-        cfg_dict["REFIMPD"] = z_terminations if z_terminations is not None else cfg_dict["REFIMPD"]
-        cfg_dict["SPECBER"] = ber if ber is not None else cfg_dict["SPECBER"]
-        cfg_dict["MODTYPE"] = modulation_type if modulation_type is not None else cfg_dict["MODTYPE"]
-        cfg_dict["FIXDELY"] = fixture_delay if fixture_delay is not None else cfg_dict["FIXDELY"]
-        cfg_dict["INPVOLT"] = input_amplitude if input_amplitude is not None else cfg_dict["INPVOLT"]
-        cfg_dict["TRSTIME"] = transition_time if transition_time is not None else cfg_dict["TRSTIME"]
-        cfg_dict["SIGBETA"] = signal_loss_factor if signal_loss_factor is not None else cfg_dict["SIGBETA"]
-        cfg_dict["REFLRHO"] = permitted_reflection if permitted_reflection is not None else cfg_dict["REFLRHO"]
-        cfg_dict["NCYCLES"] = reflections_lenght if reflections_lenght is not None else cfg_dict["NCYCLES"]
-
-        new_cfg_file = os.path.join(self.working_directory, "spisim_erl.cfg")
-        with open(new_cfg_file, "w") as fp:
-            fp.write("INPARRY = {}\n".format(cfg_dict["INPARRY"]))
-            fp.write("MIXMODE = {}\n".format(cfg_dict["MIXMODE"]))
-            fp.write("THRUS4P = {}\n".format(cfg_dict["THRUS4P"]))
-            fp.write("BANDWID = {}\n".format(cfg_dict["BANDWID"]))
-            fp.write("TDR_DUR = {}\n".format(cfg_dict["TDR_DUR"]))
-            fp.write("REFIMPD = {}\n".format(cfg_dict["REFIMPD"]))
-            fp.write("BINSIZE = {}\n".format(cfg_dict["BINSIZE"]))
-            fp.write("SPECBER = {}\n".format(cfg_dict["SPECBER"]))
-            fp.write("MODTYPE = {}\n".format(cfg_dict["MODTYPE"]))
-            fp.write("FIXDELY = {}\n".format(cfg_dict["FIXDELY"]))
-            fp.write("INPVOLT = {}\n".format(cfg_dict["INPVOLT"]))
-            fp.write("TRSTIME = {}\n".format(cfg_dict["TRSTIME"]))
-            fp.write("SIGBETA = {}\n".format(cfg_dict["SIGBETA"]))
-            fp.write("REFLRHO = {}\n".format(cfg_dict["REFLRHO"]))
-            fp.write("NCYCLES = {}\n".format(cfg_dict["NCYCLES"]))
-
-        cfgCmmd = '-i %s -v CFGFILE="%s"' % (touchstone_file, new_cfg_file)
-        command = [spisimExe, "CalcERL", cfgCmmd]
-        # Debug('%s %s' % (cmdList[0], ' '.join(arguments)))
-        # try up to three times to be sure
-        out_processing = os.path.join(self.working_directory, "spsim_erl_out.txt")
-        my_env = os.environ.copy()
-        my_env.update(settings.aedt_environment_variables)
-        if is_linux:  # pragma: no cover
-            command.append("&")
-            with open(out_processing, "w") as outfile:
-                subprocess.Popen(command, env=my_env, stdout=outfile, stderr=outfile).wait()  # nosec
-        else:
-            with open(out_processing, "w") as outfile:
-                subprocess.Popen(" ".join(command), env=my_env, stdout=outfile, stderr=outfile).wait()  # nosec
-        out_data = {}
-        try:
-            with open(out_processing, "r") as infile:
-                lines = infile.read()
-                parmDat = lines.split("[ParmDat]:", 1)[1]
-                for keyValu in parmDat.split(","):
-                    dataAry = keyValu.split("=")
-                    out_data[dataAry[0].strip()] = float(dataAry[1].strip().split()[0])
-            return out_data
-        except IndexError:
-            self.logger.error("Failed to compute ERL. Check input parameters and retry")
-            return False
-
-    @pyaedt_function_handler()
-    def compute_erl(
-        self,
-        setup_name=None,
-        config_file=None,
-        port_order=None,
-        specify_through_ports=None,
-        bandwidth=None,
-        tdr_duration=None,
-        z_terminations=None,
-        transition_time=None,
-        fixture_delay=None,
-        input_amplitude=None,
-        ber=None,
-        pdf_bin_size=None,
-        signal_loss_factor=None,
-        permitted_reflection=None,
-        reflections_lenght=None,
-        modulation_type=None,
-    ):
-        """Compute effective return loss (erl) using Ansys SPISIM.
-
-        Parameters
-        ----------
-        setup_name : str, optional
-            Name of the setup to use as the nominal. The default is
-            ``None``, in which case the active setup is used or
-            nothing is used.
-        config_file : str, optional
-            Configuration file to use as reference. The default is ``None``, in which
-            case this parameter is ignored.
-        port_order : str, optional
-            Whether to use "``EvenOdd``" or "``Incremental``" numbering for ``s4p`` files.
-            Ignored if the ports are greater than 4.
-        specify_through_ports : list, optional
-            Input and Output ports on which compute the erl. Those are ordered like ``[inp, inneg, outp, outneg]``.
-        bandwidth : float, str, optional
-            Application bandwidth: inverse of one UI (unit interval). Can be a float or str with unit ("m", "g").
-            Default is ``30e9``.
-        tdr_duration : float, optional
-            TDR duration (in second): How long the TDR tailed data should be applied. Default is ``5``.
-        z_terminations : float, optional
-            Z-Terminations: termination (Z11 and Z22) when TDR is calculated. Default is ``50``.
-        transition_time : float, str, optional
-            Transition time: how fast (slew rate) input pulse transit from 0 to Vcc volt. Default is "``10p``".
-        fixture_delay : float, optional
-            Fixture delay: delay when input starts transition from 0 to Vcc. Default is ``500e-12``.
-        input_amplitude : float, optional
-            Input amplitude: Vcc volt of step input. Default is ``1.0``.
-        ber : float, optional
-            Specified BER: At what threshold ERL is calculated. Default is ``1e-4``.
-        pdf_bin_size : float, optional
-            PDF bin size: how to quantize the superimposed value. Default is ``1e-5``.
-        signal_loss_factor : float, optional
-            Signal loss factor (Beta). See SPISIM Help for info. Default is ``1.7e9``.
-        permitted_reflection : float, optional
-            Permitted reflection (Rho). See SPISIM Help for info. Default is ``0.18``.
-        reflections_lenght : float, optional
-            Length of the reflections: how many UI will be used to calculate ERL. Default is ``1000``.
-        modulation_type : str, optional
-           Modulations type: signal modulation type "``NRZ``" or "``PAM4``". Default is "``NRZ``".
-
-        Returns
-        -------
-        bool or dict
-            A dictionary with the result from the spisimExe command, ``False`` when failed.
-        """
-
-        if not setup_name:
-            setup_name = self.nominal_sweep
-        tc_file = self.export_touchstone(
-            setup_name,
-        )
-
-        return self.compute_erl_from_touchstone(
-            tc_file,
-            config_file,
-            port_order,
-            specify_through_ports,
-            bandwidth,
-            tdr_duration,
-            z_terminations,
-            transition_time,
-            fixture_delay,
-            input_amplitude,
-            ber,
-            pdf_bin_size,
-            signal_loss_factor,
-            permitted_reflection,
-            reflections_lenght,
-            modulation_type,
-        )
